Refactor package uploader and remove unused code
