+++ sdk/jamfpro/jamfproapi_jcds2.go
@@ -10,8 +10,6 @@ package jamfpro
 import (
 	"context"
 	"fmt"
-	"io"
-	"os"
 	"path/filepath"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
@@ -19,6 +17,7 @@ import (
 	"github.com/aws/aws-sdk-go-v2/credentials"
 	"github.com/aws/aws-sdk-go-v2/feature/s3/manager"
 	"github.com/aws/aws-sdk-go-v2/service/s3"
+	"github.com/deploymenttheory/go-api-sdk-jamfpro/sdk/helpers"
 )
 
 const uriJCDS2 = "/api/v1/jcds"
@@ -53,14 +52,6 @@ type JCDS2Properties struct {
 	JCDS2Enabled              bool `json:"jcds2Enabled"`
 	FileStreamEndpointEnabled bool `json:"fileStreamEndpointEnabled"`
 	MaxChunkSize              int  `json:"maxChunkSize"`
-}
-
-// progressReader is a wrapper around an io.Reader that reports progress in kilobytes and megabytes.
-type progressReader struct {
-	reader     io.Reader
-	totalBytes int64
-	readBytes  int64
-	progressFn func(int64, int64, string) // function to call to report progress
 }
 
 // CRUD
@@ -130,7 +121,7 @@ func (c *Client) RenewJCDS2Credentials() (*ResponseJCDS2UploadCredentials, error
 	return &out, nil
 }
 
-// CreateJCDS2PackageV2 creates a new file in JCDS 2.0 using AWS SDK v2
+// CreateJCDS2PackageV2 creates a new file in JCDS 2.0 using AWS SDK v2 without creating package metadata in Jamf Pro.
 func (c *Client) CreateJCDS2PackageV2(filePath string) (*ResponseJCDS2File, error) {
 	// Step 1: Obtain AWS credentials for the package upload endpoint
 	var uploadCredentials ResponseJCDS2UploadCredentials
@@ -162,33 +153,26 @@ func (c *Client) CreateJCDS2PackageV2(filePath string) (*ResponseJCDS2File, erro
 	// Step 3: Create an Uploader with the configuration and default options
 	uploader := manager.NewUploader(s3Client)
 
-	// Open the file and use a progressReader to track the upload progress
-	file, err := os.Open(filePath)
+	// Step 3: Use the secure file reading helper
+	fileReader, fileSize, err := helpers.ReadJCDSPackageTypes(filePath)
 	if err != nil {
-		return nil, fmt.Errorf("failed to open file: %v", err)
-	}
-	defer file.Close()
-
-	fileInfo, err := file.Stat()
-	if err != nil {
-		return nil, fmt.Errorf("failed to get file info: %v", err)
+		return nil, fmt.Errorf("failed to read package file securely: %v", err)
 	}
 
-	progressFn := func(read, total int64, unit string) {
-		fmt.Printf("\rUploaded %d / %d %s (%.2f%%)", read, total, unit, float64(read)/float64(total)*100)
-	}
-
-	reader := &progressReader{
-		reader:     file,
-		totalBytes: fileInfo.Size(),
-		progressFn: progressFn,
+	// Create a progress reader
+	progressReader := &ProgressReader{
+		reader:     fileReader,
+		totalBytes: fileSize,
+		progressFn: func(read, total int64, unit string) {
+			fmt.Printf("\rUploaded %d / %d %s (%.2f%%)", read, total, unit, float64(read)/float64(total)*100)
+		},
 	}
 
 	// Create the upload input
 	uploadInput := &s3.PutObjectInput{
 		Bucket: aws.String(uploadCredentials.BucketName),
 		Key:    aws.String(uploadCredentials.Path + filepath.Base(filePath)),
-		Body:   reader,
+		Body:   progressReader,
 	}
 
 	// Step 4. Perform the upload
@@ -199,33 +183,12 @@ func (c *Client) CreateJCDS2PackageV2(filePath string) (*ResponseJCDS2File, erro
 
 	fmt.Println("\nUpload completed Successfully")
 
-	// Combine JCDS file URI and Jamf Pro package creation response for the final response
+	// Construct the final file upload response
 	finalResponse := &ResponseJCDS2File{
 		URI: fmt.Sprintf("s3://%s/%s%s", uploadCredentials.BucketName, uploadCredentials.Path, filepath.Base(filePath)),
 	}
 
 	return finalResponse, nil
-}
-
-// Read implements the io.Reader interface for progressReader, reporting upload progress in kilobytes and megabytes.
-func (r *progressReader) Read(p []byte) (int, error) {
-	n, err := r.reader.Read(p)
-	r.readBytes += int64(n)
-
-	// Report progress in more human-readable units (KB or MB)
-	const kb = 1024
-	const mb = 1024 * kb
-	readKB := r.readBytes / kb
-	totalKB := r.totalBytes / kb
-	if totalKB > kb { // If the total size is larger than 1 MB, report in MB
-		readMB := r.readBytes / mb
-		totalMB := r.totalBytes / mb
-		r.progressFn(readMB, totalMB, "MB")
-	} else { // For smaller files, report in KB
-		r.progressFn(readKB, totalKB, "KB")
-	}
-
-	return n, err
 }
 
 // DeleteJCDS2PackageV2 deletes an existing file from JCDS 2.0 using AWS SDK v2.
